--- a/azure-pipelines.yml
+++ b/azure-pipelines.yml
@@ -41,6 +41,14 @@ steps:
       npm run perf-test
     displayName: perf test
 
+  - task: GithubPRComment@0
+    displayName: 'Post Perf Results to Github Pull Request'
+    inputs:
+      githubOwner: OfficeDev
+      githubRepo: 'office-ui-fabric-react'
+      blobFilePath: '$(PerfComment.FilePath)'
+      status: '$(PerfComment.Status)'
+
   - task: GithubPRStatus@0
     displayName: 'Update Github Pull Request Status'
     inputs: