--- a/azure-pipelines.yml
+++ b/azure-pipelines.yml
@@ -46,8 +46,8 @@ steps:
     inputs:
       githubOwner: OfficeDev
       githubRepo: 'office-ui-fabric-react'
-      blobFilePath: '$(Build.SourcesDirectory)$(PerfComment.FilePath)'
-      status: '$(PerfComment.Status)'
+      blobFilePath: '$(Build.SourcesDirectory)/$(PerfCommentFilePath)'
+      status: '$(PerfCommentStatus)'
 
   - task: GithubPRStatus@0
     displayName: 'Update Github Pull Request Status'