@KevinTCoughlin 2019-04-11
cc: @aftab-hassan the size-auditor results in https://github.com/OfficeDev/office-ui-fabric-react/pull/8687#issuecomment-481916695 are surprising