@apneer 2019-04-12
Component Perf Results:
  | Scenario Name | Current Avg Total | New Avg Total | Current Avg Per Item | New Avg Per Item |
  |----------|-------------------|---------------|----------------------|------------------|
| PrimaryButtonms | 99.403ms | 102.401ms| 1.024ms | 1.024ms|
| BaseButtonms | 35.796ms | 37.400ms| 0.374ms | 0.374ms|
| NewButtonms | 108.902ms | 111.110ms| 1.111ms | 1.111ms|
| buttonms | 5.335ms | 5.075ms| 0.051ms | 0.051ms|
| DetailsRowswithoutstylesms | 204.346ms | 192.292ms| 1.923ms | 1.923ms|
| DetailsRowsms | 209.453ms | 211.280ms| 2.113ms | 2.113ms|
| Togglesms | 51.229ms | 53.001ms| 0.530ms | 0.530ms|
| NewTogglems | 68.609ms | 69.057ms| 0.691ms | 0.691ms|