@apneer 2019-04-12
Component Perf Results:
  <table>
  <tr>
    <th>Scenario</th>
    <th>Current Avg Total</th>
    <th>New Avg Total</th>
    <th>Current Avg Per Item</th>
    <th>New Avg Per Item</th>
  </tr><tr>
            <td>PrimaryButton</td>
            <td>95.823 ms</td>
            <td>95.728 ms</td>
            <td>0.958 ms</td>
            <td>0.957 ms</td>
           </tr>  
<tr>
            <td>BaseButton</td>
            <td>32.202 ms</td>
            <td>33.405 ms</td>
            <td>0.322 ms</td>
            <td>0.334 ms</td>
           </tr>  
<tr>
            <td>NewButton</td>
            <td>104.862 ms</td>
            <td>103.312 ms</td>
            <td>1.048 ms</td>
            <td>1.033 ms</td>
           </tr>  
<tr>
            <td>button</td>
            <td>4.276 ms</td>
            <td>4.829 ms</td>
            <td>0.043 ms</td>
            <td>0.048 ms</td>
           </tr>  
<tr>
            <td>DetailsRows without styles</td>
            <td>181.441 ms</td>
            <td>179.527 ms</td>
            <td>1.814 ms</td>
            <td>1.795 ms</td>
           </tr>  
<tr>
            <td>DetailsRows</td>
            <td>192.257 ms</td>
            <td>192.459 ms</td>
            <td>1.923 ms</td>
            <td>1.925 ms</td>
           </tr>  
<tr>
            <td>Toggles</td>
            <td>48.746 ms</td>
            <td>47.548 ms</td>
            <td>0.487 ms</td>
            <td>0.475 ms</td>
           </tr>  
<tr>
            <td>NewToggle</td>
            <td>68.914 ms</td>
            <td>68.041 ms</td>
            <td>0.689 ms</td>
            <td>0.681 ms</td>
           </tr>  </table>