@apneer 2019-04-12
Component Perf Results:
  <table>
  <tr>
    <th>Scenario</th>
    <th>Current Avg Total (ms)</th>
    <th>New Avg Total (ms)</th>
    <th>Current Avg Per Item (ms)</th>
    <th>New Avg Per Item (ms)</th>
  </tr><tr>
            <td>PrimaryButton</td>
            <td>103.927</td>
            <td>103.365</td>
            <td>1.039</td>
            <td>1.034</td>
           </tr>  
<tr>
            <td>BaseButton</td>
            <td>37.931</td>
            <td>36.584</td>
            <td>0.379</td>
            <td>0.366</td>
           </tr>  
<tr>
            <td>NewButton</td>
            <td>110.205</td>
            <td>111.534</td>
            <td>1.102</td>
            <td>1.115</td>
           </tr>  
<tr>
            <td>button</td>
            <td>5.680</td>
            <td>5.437</td>
            <td>0.057</td>
            <td>0.054</td>
           </tr>  
<tr>
            <td>DetailsRows without styles</td>
            <td>230.986</td>
            <td>192.900</td>
            <td>2.310</td>
            <td>1.929</td>
           </tr>  
<tr>
            <td>DetailsRows</td>
            <td>203.456</td>
            <td>210.123</td>
            <td>2.035</td>
            <td>2.101</td>
           </tr>  
<tr>
            <td>Toggles</td>
            <td>50.868</td>
            <td>51.666</td>
            <td>0.509</td>
            <td>0.517</td>
           </tr>  
<tr>
            <td>NewToggle</td>
            <td>70.076</td>
            <td>71.678</td>
            <td>0.701</td>
            <td>0.717</td>
           </tr>  </table>