@ecraig12345 2019-04-12
This is cool! How hard would it be to also do up/down indicators next to the new average values? (Probably before the value so they line up vertically.) Some possible indicators for up/down/no change:
▲ ▼ - (triangles are unicode 0x25B2 0x25BC)
:red_circle: :white_check_mark: :heavy_minus_sign: (`:red_circle: :white_check_mark: :heavy_minus_sign:`)