@apneer 2019-04-12
Component Perf Results:
  <table>
  <tr>
    <th>Scenario</th>
    <th>Current Avg Total (ms)</th>
    <th>New Avg Total (ms)</th>
    <th>Current Avg Per Item (ms)</th>
    <th>New Avg Per Item (ms)</th>
  </tr><tr>
            <td>PrimaryButton</td>
            <td>106.071</td>
            <td>106.838</td>
            <td>1.061</td>
            <td>1.068</td>
           </tr>  
<tr>
            <td>BaseButton</td>
            <td>36.294</td>
            <td>36.514</td>
            <td>0.363</td>
            <td>0.365</td>
           </tr>  
<tr>
            <td>NewButton</td>
            <td>111.712</td>
            <td>113.922</td>
            <td>1.117</td>
            <td>1.139</td>
           </tr>  
<tr>
            <td>button</td>
            <td>5.242</td>
            <td>6.040</td>
            <td>0.052</td>
            <td>0.060</td>
           </tr>  
<tr>
            <td>DetailsRows without styles</td>
            <td>200.861</td>
            <td>194.986</td>
            <td>2.009</td>
            <td>1.950</td>
           </tr>  
<tr>
            <td>DetailsRows</td>
            <td>211.405</td>
            <td>211.040</td>
            <td>2.114</td>
            <td>2.111</td>
           </tr>  
<tr>
            <td>Toggles</td>
            <td>51.441</td>
            <td>53.622</td>
            <td>0.514</td>
            <td>0.536</td>
           </tr>  
<tr>
            <td>NewToggle</td>
            <td>72.293</td>
            <td>72.431</td>
            <td>0.723</td>
            <td>0.724</td>
           </tr>  </table><p style="display:none;">perfComment9423</p>