@apneer 2019-04-12
> This is cool! How hard would it be to also do up/down indicators next to the new average values? (Probably before the value so they line up vertically.) That would make it much easier to scan. Some possible indicators for up/down/no change:
> ▲ ▼ - (triangles are unicode 0x25B2 0x25BC)
> 🔴 ✅ ➖ (`:red_circle: :white_check_mark: :heavy_minus_sign:`)

It would not be hard. For right now let's see how numbers behavior. One decision we need to make is what 'up' or 'down' means since these numbers are approximate. Once we determine what is good indication of change then we'd do that.